Calculations now significantly faster
Changed acceleration calculation stage to use a single for loop and matrix calculations, rather than having two FOR loops running and calculating the accelerations. This method significantly reduces the calculation time. In fact, simulating 100 objects for 2000 time steps takes roughly 70 seconds for the dictionary version, while this version completes the calculations in under 12 seconds. Further improvements may be had by removing the FOR loop in the calculation stage entirely, and substituting it for substantially faster numpy matrix operations.